support for custom comp in sort, reorg API/impl
